+++ src/Screens/Auth/Auth.tsx
@@ -9,75 +9,50 @@ import {
 } from 'react-native';
 import type {StackParamsList} from '../../types/stackParamList';
 import {LinearGradient} from 'expo-linear-gradient';
-import {postToken} from '../../APIs/token';
-
-import {KakaoOAuthToken, login} from '@react-native-seoul/kakao-login';
-
 import useStore from '../../Store/store';
 import {useState} from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import {logIn} from '../../APIs/member';
+import {RegisterToken} from '../../types/types';
+import {useKakaoLogin} from '../../Hooks/Auth/useKaKaoLogin';
 
 type Props = NativeStackScreenProps<StackParamsList, 'Auth'>;
 
 export function Auth({navigation}: Props) {
   const store = useStore();
 
-  const [disable, setDisable] = useState(false);
+  const [disableSignUp, setDisableSignUp] = useState(false);
 
-  function signUpWithSocialService(socialService: string) {
-    switch (socialService) {
-      case 'google':
-        navigation.navigate('NicknameForm');
-        break;
-      default:
-        break;
-    }
-  }
+  const {signUpWithKakao} = useKakaoLogin();
 
-  function isKakaoOAuthToken(arg: any): arg is KakaoOAuthToken {
-    return arg.idToken !== undefined;
-  }
-
-  const signUpWithKakao = async (): Promise<void> => {
-    setDisable(true);
+  const onPressSignUpWithKaKao = async () => {
+    setDisableSignUp(true);
     try {
-      // 카카오 로그인 호출
-      const token = await login();
-      if (isKakaoOAuthToken(token)) {
-        const tokenInfo = {
-          idToken: token.idToken,
-        };
-
-        // registerToken 발급
-        const userTokens = await postToken(tokenInfo);
-        console.log(userTokens);
-
-        if (userTokens.verified === false) {
-          store.setRegisterToken(userTokens.registerToken);
-          navigation.navigate('NicknameForm');
-        } else {
-          AsyncStorage.setItem('accessToken', userTokens.accessToken);
-          AsyncStorage.setItem('refreshToken', userTokens.refreshToken);
-
-          const userInfo = await logIn();
-          store.setUserInfo({
-            nickname: userInfo.nickname,
-            personalityIds: userInfo.personalityIds,
-            topicIds: userInfo.topicIds,
-            geolocationId: userInfo.geolocationId,
-            parentGeolocationId: userInfo.parentGeolocationId,
-          });
-
-          store.setIsLoggedIn(true);
-        }
-      } else {
-        throw new Error('KakaoOAuthToken 형식이 아님');
-      }
-    } catch (err: any) {
-      console.error(err.message);
+      const userTokens = await signUpWithKakao();
+      await loginWithToken(userTokens);
+    } catch (error: any) {
+      console.error(error.message);
     } finally {
-      setDisable(false);
+      setDisableSignUp(false);
+    }
+  };
+
+  const loginWithToken = async (userTokens: RegisterToken) => {
+    if (userTokens.verified === false) {
+      store.setRegisterToken(userTokens.registerToken);
+      navigation.navigate('NicknameForm');
+    } else {
+      AsyncStorage.setItem('accessToken', userTokens.accessToken);
+      AsyncStorage.setItem('refreshToken', userTokens.refreshToken);
+      const userInfo = await logIn();
+      store.setUserInfo({
+        nickname: userInfo.nickname,
+        personalityIds: userInfo.personalityIds,
+        topicIds: userInfo.topicIds,
+        geolocationId: userInfo.geolocationId,
+        parentGeolocationId: userInfo.parentGeolocationId,
+      });
+      store.setIsLoggedIn(true);
     }
   };
 
@@ -93,8 +68,8 @@ export function Auth({navigation}: Props) {
         </View>
         <View style={styles.buttonWrap}>
           <TouchableWithoutFeedback
-            disabled={disable}
-            onPress={signUpWithKakao}>
+            disabled={disableSignUp}
+            onPress={onPressSignUpWithKaKao}>
             <View style={[styles.loginButton, {backgroundColor: '#F9E54C'}]}>
               <Text style={[styles.loginText]}>카카오로 시작하기</Text>
             </View>
@@ -104,8 +79,7 @@ export function Auth({navigation}: Props) {
               <Text style={[styles.loginText]}>네이버로 시작하기</Text>
             </View>
           </TouchableWithoutFeedback>
-          <TouchableWithoutFeedback
-            onPress={() => signUpWithSocialService('google')}>
+          <TouchableWithoutFeedback>
             <View
               style={[
                 styles.loginButton,
